+++ libhttp-server/HttpHeaderReader.hpp
@@ -26,8 +26,8 @@ namespace HTTP {
 		void clear();
 		bool complete();
 		int cutEndPos();
-		void append(const char * data, int size);
-		int read(const char * data, int size);
+		void append(const char * data, size_t size);
+		size_t read(const char * data, size_t size);
 		bool parse();
 		HttpHeader & getHeader();
 	};
